ARM: boards: i.MX93: Add NXP FRDM i.MX93 support
This board is a low cost i.MX93 eval board from NXP.

There is no upstream DT yet, therefore a minimal DT is added in this
patch. The kernel DT is submitted in [0].

The DDR timings are copied from the NXP yocto layer [1] for u-boot with
some small modifications to remove unused timings and DDR sizes.

[0] https://lore.kernel.org/r/20250526-fpg-nxp-imx93-frdm-v2-0-e5ad0efaec33@pengutronix.de
[1] https://github.com/nxp-imx-support/meta-imx-frdm/blob/lf-6.6.36-2.1.0/meta-imx-bsp/recipes-bsp/u-boot/u-boot-imx/0002-imx-imx93_frdm-Add-basic-board-support.patch

Signed-off-by: Fabian Pflug <f.pflug@pengutronix.de>